requirements.txt
Flask
torch
torchvision
Pillow
clip-by-openai
